fixes
ok, so today has basically been a BUNCH fixes of the old so that it works with the new information. today I have re done the code for the front page(looks the same), fixed issues with the header, fixing grid(both on homepage and in new page for asignment 2). and so on